% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afghan.life.R
\docType{data}
\name{afghan.life}
\alias{afghan.life}
\title{Life table values for Afghanistan, 1950-2100.}
\format{An array with dimensions "age", "sex", "time", and "fun".}
\source{
United Nations, Department of Economic and Social Affairs,
Population Division (2015). World Population Prospects: The 2015 Revision.
Downloaded 27 September 2016.
}
\usage{
afghan.life
}
\description{
Life tables for Afghanistan constructed as part of the 2015 World Population
Prospects.  Includes the life table functions mx, qx, px, lx, dx, Lx,
Sx, Tx, ex, and ax.
}
\keyword{datasets}

