% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambodia.conc.census_census.ag.R
\docType{data}
\name{cambodia.conc.census_census.ag}
\alias{cambodia.conc.census_census.ag}
\title{Data frame with concordance between Cambodian Census regions.}
\format{A data frame, with two columns:
\describe{
  \item{census}{All 24 provinces used in 2008 Cambodian census}.
  \item{census.ag}{The same provinces, but aggregated in the
same was as the 2010 Demographic and Health Survey.}
}}
\source{
Online database of 2008 census data at
\url{celade.cepal.org/redkhm/census/khm2008}
and dataset for 2008 Cambodian DHS, obtained from
\url{dhsprogram.com}
}
\usage{
cambodia.conc.census_census.ag
}
\description{
Data frame with concordance between Cambodian Census regions.
}
\seealso{
\code{\link{cambodia.conc.dhs_census.ag}}
}
\keyword{datasets}
