% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambodia.conc.dhs_census.ag.R
\docType{data}
\name{cambodia.conc.dhs_census.ag}
\alias{cambodia.conc.dhs_census.ag}
\title{Data frame with concordance between Cambodian DHS and Census regions.}
\format{A data frame, with two columns:
\describe{
  \item{dhs}{Regions used in 2010 Demographic and Health Survey for Cambodia.
Consists of 14 individual provinces and 5 pairs of provinces}
  \item{census.ag}{The same regions, but with the spellings used in the
2008 Cambodian census.}
}}
\source{
Online database of 2008 census data at
\url{celade.cepal.org/redkhm/census/khm2008}
and dataset for 2008 Cambodian DHS, obtained from
\url{dhsprogram.com}
}
\usage{
cambodia.conc.dhs_census.ag
}
\description{
Data frame with concordance between Cambodian DHS and Census regions.
}
\seealso{
\code{\link{cambodia.conc.census_census.ag}}
}
\keyword{datasets}
