% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cambodia.popn.R
\docType{data}
\name{cambodia.popn}
\alias{cambodia.popn}
\title{Estimates of Cambodian population from 2008 census.}
\format{An array with dimensions "region" (province), "area" (urban-rural)
"age",  and "sex".}
\source{
Online database of 2008 census data at
\url{http://celade.cepal.org/redkhm/census/khm2008}
}
\usage{
cambodia.popn
}
\description{
Estimates of Cambodian population from 2008 census.
}
\keyword{datasets}
