% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demdata-package.R
\docType{package}
\name{demdata}
\alias{demdata}
\alias{demdata-package}
\title{demdata: Example demographic data.}
\description{
The datasets are used  by packages 'demdata', 'demest',
'demlife', 'demfam', and 'classconc'.  Most datasets are arrays,
though some are data.frames.  Most of the data is real,
but some is simulated.
}

