% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutch.cpi.R
\docType{data}
\name{dutch.cpi}
\alias{dutch.cpi}
\title{Consumer Price Index in the Netherlands, 1996-2015.}
\format{A data.frame with columns 'time' and 'value'.}
\source{
Table "Key Short-Term Economic Indicators: Consumer
Prices - Annual inflation"from OECD database 'OECD.Stat'.
Data downloaded 3 June 2016.
}
\usage{
dutch.cpi
}
\description{
Consumer Price Index in the Netherlands, 1996-2015.
}
\keyword{datasets}
