% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutch.popn.R
\docType{data}
\name{dutch.popn}
\alias{dutch.popn}
\title{Estimated and projected population of the Netherlands,
by age and sex, 1950-2050.}
\format{An array with dimensions "age", "sex", and "time".}
\source{
Table "Historical population data and projections
(1950-2050)" from OECD database 'OECD.Stat'.
Data downloaded 5 June 2016.
}
\usage{
dutch.popn
}
\description{
Estimated and projected population of the Netherlands,
by age and sex, 1950-2050.
}
\keyword{datasets}

