% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/england.wales.popn.R
\docType{data}
\name{england.wales.popn}
\alias{england.wales.popn}
\title{Mid-Year Population in England and Wales in 2014 by Age, Sex,
and Local Administrative District (LAD).}
\format{An array with dimensions "age", "sex", and "lad".}
\source{
Dataset
"MYE2_population_by_sex_and_age_for_local_authorities_UK_2014"
downloaded from the ONS website on 21 October 2016.
}
\usage{
england.wales.popn
}
\description{
The data are from the Office for National Statistics website.
}
\keyword{datasets}

