% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceland.births.R
\docType{data}
\name{iceland.births}
\alias{iceland.births}
\title{Births in Iceland, by Age of Mother and Year}
\format{An array with dimensions "age" and "year".}
\source{
Database "Live births by age of mother 1971-2015".
Downloaded from Statistics Iceland website on 22 January 2017.
}
\usage{
iceland.births
}
\description{
The data are from the Statistics Iceland website.
}
\keyword{datasets}
