% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceland.migrants.R
\docType{data}
\name{iceland.migrants}
\alias{iceland.migrants}
\title{Counts of International Migrants to and from Iceland}
\format{An array with dimensions "age", "sex", "year", "citizenship",
and "direction" (ie immigrants vs emigrants).}
\source{
Database "External migration by sex, age and citizenship 1986-2015".
Downloaded from Statistics Iceland website on 21 January 2017.
}
\usage{
iceland.migrants
}
\description{
The data are from the Statistics Iceland website.  The data contain counts
of immigrants and emigrants, and distinguish between Icelandic citizens
and foreigners.
}
\keyword{datasets}
