% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceland.popn.R
\docType{data}
\name{iceland.popn}
\alias{iceland.popn}
\title{Population of Iceland, by Age, Sex, and Year}
\format{An array with dimensions "age", "sex", and "year".}
\source{
Database "Population by sex and age 1841-2016".
Downloaded from Statistics Iceland website on 21 January 2017.
}
\usage{
iceland.popn
}
\description{
The data are from the Statistics Iceland website.
}
\keyword{datasets}

