% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/india.fert.R
\docType{data}
\name{india.fert}
\alias{india.fert}
\title{Estimates of Indian fertility rates (per thousand) by age of mother and period.}
\format{An array with dimensions "age" and "period".}
\source{
United Nations, Department of Economic and Social Affairs,
Population Division (2015). World Population Prospects: The 2015 Revision,
custom data acquired via website.
}
\usage{
india.fert
}
\description{
Estimates of Indian fertility rates (per thousand) by age of mother and period.
}
\keyword{datasets}

