% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/india.popn.R
\docType{data}
\name{india.popn}
\alias{india.popn}
\title{Estimates of Indian population (thousands) by age and sex, 2010.}
\format{An array with dimensions "age" and "sex".}
\source{
United Nations, Department of Economic and Social Affairs,
Population Division (2015). World Population Prospects: The 2015 Revision,
custom data acquired via website.
}
\usage{
india.popn
}
\description{
Estimates of Indian population (thousands) by age and sex, 2010.
}
\keyword{datasets}

