% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maori.deaths.R
\docType{data}
\name{maori.deaths}
\alias{maori.deaths}
\title{Deaths of Maori in New Zealand.}
\format{An array with dimensions \code{"age"}, \code{"sex"}, and \code{"year"}.}
\source{
Statistics New Zealand website, downloaded 24 October 2016.
}
\usage{
maori.deaths
}
\description{
Counts of registered deaths for Maori, the indigenous people of New Zealand.
}
\details{
The years are "year to December", ie 1 January to 31 December.

The data have been confidentialised by Statistics New Zealand by rounding
randomly to base 3.
}
\keyword{datasets}
