% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maori.popn.R
\docType{data}
\name{maori.popn}
\alias{maori.popn}
\title{Population of Maori in New Zealand.}
\format{An array with dimensions \code{"age"}, \code{"sex"}, and
\code{"year"}.}
\source{
Statistics New Zealand website, downloaded 24 October 2016.
}
\usage{
maori.popn
}
\description{
Population estimates for Maori, the indigenous people of New Zealand.
}
\details{
The estimates are 'mid-year' estimates, ie they give the estimated
population on 30 June.

The data have been confidentialised by Statistics New Zealand by rounding
to the nearest 10.
}
\keyword{datasets}

