% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.arrivals.R
\docType{data}
\name{nz.arrivals}
\alias{nz.arrivals}
\title{Permament and long-term arrivals in New Zealand, 1991-2015.}
\format{An array with dimensions "region", "age", "sex", and "year".}
\source{
Calculated from data on Statistics New Zealand website.
Downloaded 21 July 2016.
}
\usage{
nz.arrivals
}
\description{
A permanent and long-term arrival is an arrival in New Zealand
of a person who intends to stay in the country for at least one year.
}
\details{
The data are disaggregated by age, sex, calendar year, and
territorial authority.  The territorial authority is the
most important sub-national administrative unit in
New Zealand.

In 2010, seven territorial authorities were
amalgamated to form a single Auckland region.  \code{nz.arrivals}
uses the post-amalgamation boundaries for the entire
period 1991-2015.
}
\seealso{
\code{\link{nz.departures}}, \code{\link{nz.popn.ta}}
}
\keyword{datasets}

