% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.births.R
\docType{data}
\name{nz.births}
\alias{nz.births}
\title{Registered births in New Zealand by age, sex, and region, 1991-2015.}
\format{An array with dimensions "age", "region", and "year"}
\source{
Statistics New Zealand website.  Downloaded 9 June 2016.
}
\usage{
nz.births
}
\description{
The age group \code{"Under 20 Years"} in the original data has been
renamed \code{"15-19"}, and the age group \code{"40 Years and Over"} has
been renamed \code{"40-44"}.
}
\keyword{datasets}

