% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.deaths.reg.R
\docType{data}
\name{nz.deaths.reg}
\alias{nz.deaths.reg}
\title{Registered deaths in New Zealand by age, sex, and region, 1997-2016.}
\format{An array with dimensions "age", "sex", "region", and "year"}
\source{
Custom tabulation from Statistics New Zealand.
}
\usage{
nz.deaths.reg
}
\description{
The years are June years.  For instance, the year "1997" runs from 1 July 1996
to 30 Jun 1997.
}
\seealso{
\code{\link{nz.popn.reg}}, \code{\link{nz.births.reg}},
\code{\link{nz.int.mig.reg}}, \code{\link{nz.ext.mig.reg}}.
}
\keyword{datasets}
