% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.income.R
\docType{data}
\name{nz.income}
\alias{nz.income}
\title{New Zealand Income Survey data.}
\format{An data.frame with 11315 rows and the following variables:
\describe{
  \item{id}{Individual-level ID.}
  \item{age}{Age, in five-year age groups.}
  \item{sex}{Female or male.}
  \item{ethnicity}{Stated ethnicity or ethnicities.}
  \item{qualification}{Highest qualifification.}
  \item{hours}{Hours worked per week.}
  \item{income}{Income per week (in NZ dollars).}
}}
\source{
Statistics New Zealand website.
}
\usage{
nz.income
}
\description{
A 'synthetic unit record file' based on the 2011 New Zealand Income Survey.
}
\keyword{datasets}
