% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.mort.R
\docType{data}
\name{nz.mort}
\alias{nz.mort}
\title{Mortality rates for New Zealand}
\format{An array with dimensions \code{"age"}, \code{"sex"}, and \code{"ethnicity"}.}
\source{
Statistics New Zealand website
}
\usage{
nz.mort
}
\description{
NEED TO REVISE THIS AND DO DOCUMENTATION
}
\keyword{datasets}
