% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.obesity.mean.R
\docType{data}
\name{nz.obesity.mean}
\alias{nz.obesity.mean}
\title{Prevalence of obesity in New Zealand.}
\format{An matrix with age, in 10-year age groups, by time.  The time
dimension includes years in which there was no survey.}
\source{
New Zealand Ministry of Health website.
}
\usage{
nz.obesity.mean
}
\description{
Estimates from New Zealand Health Survey, 1997, 2003, 2007, 2012, 2013.
}
\seealso{
Standard errors for the estimates are in \code{\link{nz.obesity.se}}
}
\keyword{datasets}

