% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.obesity.se.R
\docType{data}
\name{nz.obesity.se}
\alias{nz.obesity.se}
\title{Standard errors for estimates of prevalence of obesity in New Zealand.}
\format{An matrix with age, in 10-year age groups, by time.  The time
dimension includes years in which there was no survey.}
\source{
Calculated from reports on the New Zealand Ministry of Health website.
}
\usage{
nz.obesity.se
}
\description{
The standard errors were calculated from confidence intervals in
reports on the New Zealand Health Survey, 1997, 2003, 2007, 2012, 2013.
}
\seealso{
Estimates are in \code{\link{nz.obesity.mean}}
}
\keyword{datasets}
