% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.popn.reg.R
\docType{data}
\name{nz.popn.reg}
\alias{nz.popn.reg}
\title{Estimates of New Zealand population by age, sex, and region, 2006-2014.}
\format{An array with dimensions "age", "sex", "region", and "year"}
\source{
Statistics New Zealand website.
}
\usage{
nz.popn.reg
}
\description{
Estimates of New Zealand population by age, sex, and region, 2006-2014.
}
\keyword{datasets}

