% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.popn.ta.R
\docType{data}
\name{nz.popn.ta}
\alias{nz.popn.ta}
\title{Estimates of New Zealand population by age, sex,
and territorial authority, 30 June, 1996-2015.}
\format{An array with dimensions "age", "sex", "region", and "year"}
\source{
Data for 1996, 2001, and 2006-2015 downloaded from
   Statistics New Zealand website, 21 July 2016.  Values
   for 1997-2000 and 2002-2005 obtained by spline interpolation.
}
\usage{
nz.popn.ta
}
\description{
The territorial authority is the most important
sub-national administrative unit in New Zealand.
}
\details{
In 2010, seven territorial authorities were
amalgamated to form a single Auckland region.  \code{nz.popn.ta}
uses the post-amalgamation boundaries for the entire
period 1996-2015.
}
\keyword{datasets}

