% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.school.R
\docType{data}
\name{nz.school}
\alias{nz.school}
\title{Number of New Zealanders enrolled in school, by age, sex, and region, 1996-2016.}
\format{An array with dimensions "age", "sex", "region", and "year".}
\source{
Downloaded from New Zealand Ministry of Education website,
\url{https://www.educationcounts.govt.nz/statistics/schooling/student-numbers/6028}
on 23 Jun 2017.
}
\usage{
nz.school
}
\description{
Numbers of New Zealanders aged 5 and above, enrolled in school, by
age, sex, and regional council area, on 1 July of each year.
}
\details{
Students in the Chatham Islands have been recoded to the Canterbury region.
}
\keyword{datasets}

