% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.visitors.R
\docType{data}
\name{nz.visitors}
\alias{nz.visitors}
\title{Number of visitors coming to New Zealand for a holiday, by quarter, 1978Q2-2016Q2.}
\format{An array with dimensions "age", "sex", and "time".}
\source{
Statistics New Zealand website. Downloaded 7 September 2016.
}
\usage{
nz.visitors
}
\description{
The data refer to the 8 main source countries for visits to New Zealand:
Australia, Canada, China, Germany, Japan, Korea, the UK, and the US.
}
\details{
The data are disaggregated by age, sex, quarter, and country of origin.
}
\seealso{
\code{\link{nz.intl.migr}} has data on 'permanent and long-term'
arrivals and departures in New Zealand.
}
\keyword{datasets}
