% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/russia.births.R
\docType{data}
\name{russia.births}
\alias{russia.births}
\title{Births in the Russian Federation by Sex of Child, Age of Mother, and Year.}
\format{An data.frame with columns "year", "sex", "age.of.mother", and "count".}
\source{
United Nations Statistics Division online database 'UNSD Demographic Statistics'.  Downloaded 5 March 2016.
}
\usage{
russia.births
}
\description{
The data come from birth registrations, but refer to the year the birth was occurred, not the year it was registered.
}
\keyword{datasets}
