% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.admin.health.R
\docType{data}
\name{sim.admin.health}
\alias{sim.admin.health}
\title{Simulated health enrolment data.}
\format{An array with dimensions \code{"age"}, \code{"sex"}, \code{"region"},
and \code{"year"}.}
\usage{
sim.admin.health
}
\description{
The data set is generated by scaling population counts for New Zealand by
coverage rates that vary by age.
}
\seealso{
The other simulated administrative datasets are
\code{\link{sim.admin.nat}}, \code{\link{sim.admin.school}},
and \code{\link{sim.admin.survey}}.  The original New Zealand population
dataset is \code{\link{nz.popn.ta}}.
}
\keyword{datasets}

