% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.admin.nat.R
\docType{data}
\name{sim.admin.nat}
\alias{sim.admin.nat}
\title{Simulated national-level administrative data.}
\format{An array with dimensions \code{"age"}, \code{"sex"}, and \code{"year"}.}
\usage{
sim.admin.nat
}
\description{
The dataset was constructed by aggregating over the \code{region} dimension
of New Zealand population data, and then adding a small amount of random
noise.
}
\seealso{
The other simulated administrative datasets are
\code{\link{sim.admin.health}}, \code{\link{sim.admin.school}},
and \code{\link{sim.admin.survey}}.  The original New Zealand population
dataset is \code{\link{nz.popn.ta}}.
}
\keyword{datasets}

