% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.admin.school.R
\docType{data}
\name{sim.admin.school}
\alias{sim.admin.school}
\title{Simulated school enrolment data.}
\format{An array with dimensions \code{"age"}, \code{"sex"}, \code{"region"},
and \code{"year"}.}
\usage{
sim.admin.school
}
\description{
The data set is constructed by subsetting taking population counts for
\code{\link{sim.popn.true}} for ages \code{"5-9"} and \code{"10-14"} and adding
some random noise.
}
\seealso{
The other simulated administrative datasets are
\code{\link{sim.admin.health}}, \code{\link{sim.admin.nat}},
and \code{\link{sim.admin.survey}}.  The underlying "simulation-true"
population counts are \code{\link{sim.popn.true}}.
}
\keyword{datasets}

