% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.deaths.R
\docType{data}
\name{sim.deaths}
\alias{sim.deaths}
\title{Simulated counts of deaths.}
\format{An array with dimensions "age", "sex", "region", and "iteration".}
\source{
Calculated from \code{\link{us.exposure}} and
\code{\link{us.deaths}}.
}
\usage{
sim.deaths
}
\description{
One hundred simulated sets of counts of deaths by age, sex, and region.
The deaths were generated through drawing from Poisson distributions,
the expected values for which were equal to
\code{\link{sim.mortality.rate}} multiplied by \code{\link{sim.exposure}}.
}
\keyword{datasets}

