% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.exposure.R
\docType{data}
\name{sim.exposure}
\alias{sim.exposure}
\title{Simulated population exposed to risk of dying.}
\format{An array with dimensions "age", "sex", "region", and "iteration".}
\source{
Calculated from \code{\link{us.exposure}}.
}
\usage{
sim.exposure
}
\description{
One hundred simulated sets of counts of population by age, sex, and region.
Each region has an identical population, which is equal to
estimates of the population of the United States in 2013, scaled
to sum to 100,000 in each region.
}
\seealso{
\code{link{sim.deaths}}, \code{\link{sim.mortality.rate}}
}
\keyword{datasets}

