% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.mortality.rate.R
\docType{data}
\name{sim.mortality.rate}
\alias{sim.mortality.rate}
\title{Simulated mortality rates.}
\format{An array with dimensions "age", "sex", "region", and "iteration".}
\source{
Calculated from \code{\link{us.exposure}} and
\code{\link{us.deaths}}.
}
\usage{
sim.mortality.rate
}
\description{
One hundred simulated sets of death rates by age, sex, and region.
The simulated rates are based on US national-level rates for 2013.  Rates
for each simulated region were obtained by perturbing the national-level
rates and then randomly shifting the regional rates up or down by a small
amount.
}
\seealso{
\code{\link{sim.deaths}} and \code{\link{sim.exposure}}
}
\keyword{datasets}

