% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.popn.true.R
\docType{data}
\name{sim.popn.true}
\alias{sim.popn.true}
\title{Simulated population data.}
\format{An array with dimensions \code{"age"}, \code{"sex"}, \code{"region"},
and \code{"year"}.}
\usage{
sim.popn.true
}
\description{
The dataset was obtained by randomly selecting 20 regions from the New Zealand
population data \code{\link{nz.popn.ta}}.
}
\details{
The acts as the underlying counts, or "simulation truth", for datasets
\code{\link{sim.admin.nat}}, \code{\link{sim.admin.health}},
\code{\link{sim.admin.school}}, and \code{\link{sim.admin.survey}}.
}
\seealso{
The data were created from \code{\link{nz.popn.ta}}.
}
\keyword{datasets}

