% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.unemployed.R
\docType{data}
\name{sim.unemployed}
\alias{sim.unemployed}
\title{Simulated counts of unemployed by age, sex, and region.}
\format{An array with dimensions "age", "sex", "region", and "iteration".}
\source{
Modifed from "Employment status of the civilian noninstitutional population
by age, sex, and race, retrieved from http://www.bls.gov/cps/cpsaat03.htm on 25 October 2015
}
\usage{
sim.unemployed
}
\description{
One hundred simulated sets of counts of unemployed by age, sex, and region.
The rates were generated by making Poisson draws, with means equal to
\code{\link{sim.unemployment.rate}} multipled by \code{\link{sim.labor.force}}.
}
\seealso{
\code{link{sim.unemployment.rate}}, \code{\link{sim.labor.force}}
}
\keyword{datasets}

