% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.unemployment.rate.R
\docType{data}
\name{sim.unemployment.rate}
\alias{sim.unemployment.rate}
\title{Simulated unemployment rate by age, sex, and region.}
\format{An array with dimensions "age", "sex", "region", and "iteration".}
\source{
Modifed from "Employment status of the civilian noninstitutional population
by age, sex, and race, retrieved from http://www.bls.gov/cps/cpsaat03.htm on 25 October 2015
}
\usage{
sim.unemployment.rate
}
\description{
One hundred simulated sets of unemployment rates by age, sex, and region.
The rates were generated by adding random noise to national-level
unemployment rates by age and sex for the United States in 2014.
}
\seealso{
\code{link{sim.unemployed}}, \code{\link{sim.labor.force}}
}
\keyword{datasets}
