% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweden.births.R
\docType{data}
\name{sweden.births}
\alias{sweden.births}
\title{Counts of Births in Sweden by Age of Mother, Sex of Child, County, and Year.}
\format{An array with dimensions "age", "sex", "region", and "year".}
\source{
Database "Live births by region, sex and age of mother.
Year 1968 - 2015". Downloaded from Statistics Sweden on 17 February 2017.
}
\usage{
sweden.births
}
\description{
The data are from the Statistics Sweden website.
}
\details{
Accents have been and umlauts have been removed from county names
to avoid problems with encoding. (Apologies to Swedish speakers.)
}
\keyword{datasets}
