% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweden.deaths.R
\docType{data}
\name{sweden.deaths}
\alias{sweden.deaths}
\title{Counts of Deaths in Sweden by Age, Sex, County, and Year.}
\format{An array with dimensions "age", "sex", "region", and "year".}
\source{
Database "Deaths by region, age (during the year) and sex.
Year 1968 - 2015". Downloaded from Statistics Sweden on 18 January 2017.
}
\usage{
sweden.deaths
}
\description{
The data are from the Statistics Sweden website.
}
\details{
Accents have been and umlauts have been removed from county names
to avoid problems with encoding. (Apologies to Swedish speakers.)
}
\keyword{datasets}

