% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweden.hmd.deaths.R
\docType{data}
\name{sweden.hmd.deaths}
\alias{sweden.hmd.deaths}
\title{Deaths in Sweden by age, sex, and calendar year.}
\format{An array with dimensions "age", "sex" and "year".}
\source{
Human Mortality Database. University of California, Berkeley (USA),
and Max Planck Institute for Demographic Research (Germany).
Available at \url{www.mortality.org} or \url{www.humanmortality.de}
(data downloaded on 9 September 2017).
}
\usage{
sweden.hmd.deaths
}
\description{
The data are from the Human Mortality Database,
rounded to the nearest integer.  More recent versions,
and a comprehensive description of the data,
are available at the HMD website.
}
\keyword{datasets}
