% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweden.popn.R
\docType{data}
\name{sweden.popn}
\alias{sweden.popn}
\title{Counts of Population in Sweden by Age, Sex, County, and Year.}
\format{An array with dimensions "age", "sex", "region", and "year".}
\source{
Database "Population by region, marital status, age and sex.
Year 1968 - 2015". Downloaded from Statistics Sweden on 18 January 2017.
}
\usage{
sweden.popn
}
\description{
The data are from the Statistics Sweden website.  The data refer
to the population as of 31 December in each year.  For instance, the
data for 2015 refer to the population as of 31 December 2015.
}
\details{
Accents have been and umlauts have been removed from county names
to avoid problems with encoding. (Apologies to Swedish speakers.)
}
\keyword{datasets}

