% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.exposure.R
\docType{data}
\name{uk.exposure}
\alias{uk.exposure}
\title{Years of Exposure in the United Kingdom by Age, Sex, and Calendar Year.}
\format{An array with dimensions "age", "sex" and "year".}
\source{
Human Mortality Database. University of California, Berkeley (USA),
and Max Planck Institute for Demographic Research (Germany).
Available at \url{www.mortality.org} or \url{www.humanmortality.de}
(data downloaded on 19 May 2016).
}
\usage{
uk.exposure
}
\description{
The data are from the Human Mortality Database.
More recent versions, and a comprehensive description of the data,
are available at the HMD website.
}
\keyword{datasets}
