% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.sex.ratio.R
\docType{data}
\name{uk.sex.ratio}
\alias{uk.sex.ratio}
\title{Sex Ratio in the United Kingdom, 1938-2014}
\format{An data.frame with columns "year" and "sex.ratio".}
\source{
Birth Summary Tables, England and Wales 2014, UK Office for National Statistics website.
Downloaded 15 May 2016.
}
\usage{
uk.sex.ratio
}
\description{
The sex ratio is the number of male births per 100 female births.
}
\keyword{datasets}

