% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.tfr.R
\docType{data}
\name{uk.tfr}
\alias{uk.tfr}
\title{Total Fertility Rate in the United Kingdom, 1938-2014}
\format{An data.frame with columns "year" and "tfr".}
\source{
Birth Summary Tables, England and Wales 2014, UK Office for National Statistics website.
Downloaded 15 May 2016.
}
\usage{
uk.tfr
}
\description{
The total fertility rate is the number of births the average woman would
have over her life time if current age-specific fertility rates were to
continue indefinitely.
}
\keyword{datasets}

