% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us.deaths.R
\docType{data}
\name{us.deaths}
\alias{us.deaths}
\title{Deaths in the United States by age, sex, and calendar year.}
\format{An array with dimensions "age", "sex" and "year".}
\source{
Human Mortality Database. University of California, Berkeley (USA),
and Max Planck Institute for Demographic Research (Germany).
Available at \url{www.mortality.org} or \url{www.humanmortality.de}
(data downloaded on 1 December 2015).
}
\usage{
us.deaths
}
\description{
The data are from the Human Mortality Database,
rounded to the nearest integer.  More recent versions,
and a comprehensive description of the data,
are available at the HMD website.
}
\keyword{datasets}

