% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us.exposure.R
\docType{data}
\name{us.exposure}
\alias{us.exposure}
\title{Years of exposure in the United States by age, sex, and calendar year.}
\format{An array with dimensions "age", "sex" and "year".}
\source{
Human Mortality Database. University of California, Berkeley (USA),
and Max Planck Institute for Demographic Research (Germany).
Available at \url{www.mortality.org} or \url{www.humanmortality.de}
(data downloaded on 1 December 2015).
}
\usage{
us.exposure
}
\description{
The data are from the Human Mortality Database.
More recent versions, and a comprehensive description of the data,
are available at the HMD website.
}
\keyword{datasets}

