% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VADeaths2.R
\docType{data}
\name{VADeaths2}
\alias{VADeaths2}
\title{Death rates in Virginia (1940) - reformatted.}
\format{An array with dimensions \code{"age"}, \code{"sex"}, and \code{"residence"}.}
\source{
See \code{\link[datasets]{VADeaths}}
}
\usage{
VADeaths2
}
\description{
The \code{\link[datasets]{VADeaths}} data from the \pkg{datasets} package,
rearranged into a 3-dimensional array.
}
\keyword{datasets}
