% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAPopn.R
\docType{data}
\name{VAPopn}
\alias{VAPopn}
\title{Population of Virginia (1940)}
\format{An array with dimensions \code{"age"}, \code{"sex"},
 \code{"residence"}, and \code{"color"}.}
\source{
Molyneaux, L., Gilliam, S. K., and Florant, L. C.(1947) Differences in
 Virginia death rates by color, sex, age, and rural or urban
 residence. \emph{American Sociological Review}, 12, 525-535.
}
\usage{
VAPopn
}
\description{
The population of Virginia in 1940 classified by age, sex, rural-urban
 residence, and 'color'.
}
\seealso{
\code{\link{VADeaths2}}
}
\keyword{datasets}
