% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waikato.tfr.R
\docType{data}
\name{waikato.tfr}
\alias{waikato.tfr}
\title{Total fertility rates in the Waikato region of New Zealand.}
\format{An array with dimensions "year" and "iteration".}
\source{
Calculated from \code{\link{nz.births}} and
\code{\link{nz.popn.reg}}.
}
\usage{
waikato.tfr
}
\description{
The total fertility rate is the number of children that a woman would
bear over her lifetime if current age-specific fertility rates were to
continue indefinitely.
}
\details{
The estimates are for the period 2006-2014 and were constructed from
\code{\link{nz.births}} and \code{\link{nz.popn.reg}},
using function \code{\link[demest]{estimateModel}} in package
\code{demest}.  \code{waikato.tfr} contains 1,000 iterations.
Variation across the iterations captures uncertainty about the
true rates.
}
\keyword{datasets}

