% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world.popn.R
\docType{data}
\name{world.popn}
\alias{world.popn}
\title{Estimates of world population.}
\format{An array with dimensions "age", "sex", and "time"}
\source{
United Nations, Department of Economic and Social Affairs,
Population Division (2015). World Population Prospects: The 2015 Revision,
via R package \pkg{wpp2015}.
}
\usage{
world.popn
}
\description{
Estimates of world population by 5-year age group and sex, from the World Population Prospects.
}
\keyword{datasets}

