% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.injuries.R
\docType{data}
\name{nz.injuries}
\alias{nz.injuries}
\title{Non-fatal injurines in New Zealand.}
\format{An data.frame with 743 rows and the following variables:
\describe{
  \item{year}{Calendar year in which the injury occurred.}
  \item{age}{Age, in broad age groups.}
  \item{sex}{Female or male.}
  \item{cause}{Cause of injury.}
  \item{count}{Number of injuries.}
}}
\source{
Table "Count of fatal and serious non-fatal injuries by sex,
age group, ethnicity, cause, and severity of injury, 2000-15" on
Statistics New Zealand website. Downloaded 3 July 2017.
}
\usage{
nz.injuries
}
\description{
Counts of non-fatal injuries in New Zealand, by year, age, sex, and cause.
}
\keyword{datasets}
