% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.popn.reg.R
\docType{data}
\name{nz.popn.reg}
\alias{nz.popn.reg}
\title{Estimates of New Zealand population by age, sex, and region, 1996-2016.}
\format{An array with dimensions "age", "sex", "region", and "year"}
\source{
Custom tabulation from Statistics New Zealand.
}
\usage{
nz.popn.reg
}
\description{
The population counts refer to 30 June of each year.
}
\seealso{
\code{\link{nz.births.reg}}, \code{\link{nz.deaths.reg}},
\code{\link{nz.int.mig.reg}}, \code{\link{nz.ext.mig.reg}}.
}
\keyword{datasets}

