% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.admin.survey.R
\docType{data}
\name{sim.admin.survey}
\alias{sim.admin.survey}
\title{Simulated survey data.}
\format{An array with dimensions \code{"age"}, \code{"sex"}, \code{"region"},
and \code{"year"}.}
\usage{
sim.admin.survey
}
\description{
The dataset is constructed by taking selecting 5\% of the \code{\link{sim.popn.true}}
population, using function \code{\link{rbinom}}.
}
\seealso{
The other simulated administrative datasets are
\code{\link{sim.admin.health}}, \code{\link{sim.admin.nat}},
and \code{\link{sim.admin.school}}.  The underlying "simulation-true"
population counts are \code{\link{sim.popn.true}}.
}
\keyword{datasets}
